Setup pre_exec at the cell level
