+++ auraed/Cargo.toml
@@ -52,6 +52,7 @@ log = "0.4.17"
 multi_log = "0.1.2"
 netlink-packet-route = "0.13.0" # Used for netlink_packet_route::rtnl::address::nlas definition
 nix = { version = "0.26.1", features = ["sched"] }
+procfs = "0.14.2"
 rtnetlink = "0.11.0"
 simplelog = "0.12.0"
 sys-mount = "2.0.1"
